Release 0.8.3: ONNX export/inference, AMP patch inference, CustomLoss fix
- vision_patch: add OnnxModelWrapper + export_to_onnx (CPU ONNX inference)
- vision_patch: add AMP (mixed-precision) to PatchInferenceEngine and
  patch_inference; set model.eval() once; resolve device for ONNX wrappers
- vision_patch: reorient via tio.ToOrientation in _postprocess. Note:
  predict(return_probabilities=True) now returns [C, ...] (was [1, C, ...]);
  segmentation masks unchanged
- vision_loss: cast targets to float in CustomLoss for dtype consistency
- Bump version 0.8.2 -> 0.8.3; add onnx/onnxruntime to dev_requirements

+++ fastMONAI/vision_loss.py
@@ -26,7 +26,8 @@ class CustomLoss:
         
         if isinstance(pred, MedBase):
             pred, targ = torch.Tensor(pred.cpu()), torch.Tensor(targ.cpu().float())
-            
+
+        targ = targ.float()
         return self.loss_func(pred, targ)
 
     def activation(self, x):
